refactor: 설정파일 위치 변경
