+++ src/main/java/kakaotech/communityBE/config/RedisSessionConfig.java
@@ -1,0 +1,4 @@
+package kakaotech.communityBE.config;
+
+public class RedisSessionConfig {
+}
